requirements.txt
arrow==0.6.0
feedparser==5.2.1
icalendar==3.9.1
parsedatetime==1.5
python-dateutil==2.4.2
pytz==2015.4
PyYAML==3.11
requests==2.7.0
six==1.9.0
wheel==0.24.0
